python_upload_service/requirements.txt
Flask==2.3.3
pandas==2.1.1
mysql-connector-python==8.1.0
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.0.0
Flask-CORS==4.0.0